chore: fix CompositionSpriteShape bug and move temp methods to SkiaHelper
